Add finding the closest GSP to the location (#673)
#minor 

* Add finding the closest GSP to the location

* Revert some config changes from local machine

* Update tests and use built-in filler

* Remove validation NaN to 0s

Running a new one with 16000 training batches, and 7000 test/validation batches, there were no errors or NaNs in the data. I think the issue was arising from issues between the newer version of dataset and the v15 data, and updates to teh input data as well. I don't think it will be a problem for new datasets, so should not be included.

* Add unit test for close GSP